fix(test/iot): 修 CI — 新服务 contextLoads 接 Testcontainers + iot schema 对齐
evidence/water/iot 的 contextLoads 此前无测试数据源，Flyway 连不上 DB 致 CI 红：
- water: Testcontainers MySQL + flyway + ddl-auto=validate
- iot:   同上 + @MockBean 掉 Mqtt5Blocking/AsyncClient 与 MqttSubscriberService
        （MqttConfig 在 bean 初始化即同步连 broker，测试环境无 broker 会失败）
- evidence: Testcontainers + flyway off + ddl-auto=none（表由 water-service 维护）

fix(iot): V044 device_telemetry_filter.life_percent TINYINT→SMALLINT，对齐实体
TelemetrySampleFilter.lifePercent(Short)；否则 ddl-auto=validate 在生产启动即崩。
（iot 迁移未部署，直接改无 Flyway checksum 风险）

本地 Testcontainers 三服务测试均通过。

+++ iot-gateway-service/src/main/resources/db/migration/V044__device_telemetry_filter.sql
@@ -1,7 +1,7 @@
 CREATE TABLE device_telemetry_filter (
   id           BIGINT PRIMARY KEY AUTO_INCREMENT,
   sn           VARCHAR(64) NOT NULL,
-  life_percent TINYINT NOT NULL COMMENT '0-100 filter life remaining',
+  life_percent SMALLINT NOT NULL COMMENT '0-100 filter life remaining',
   sampled_at   DATETIME(3) NOT NULL,
   received_at  DATETIME(3) NOT NULL DEFAULT CURRENT_TIMESTAMP(3),
   KEY idx_sn_time (sn, sampled_at)
